# Lista 3/zad-1/src/main/java/com/example/zad1/model/Order.java
package com.example.zad1.model;

import javax.persistence.*;
import java.util.ArrayList;
import java.util.List;

@Entity(name = "Orders")
public class Order {
    @Id
    @SequenceGenerator(name = "order_sequence", sequenceName = "order_sequence", allocationSize = 1)
    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "order_sequence")
    private Long id;
    private String customer_name;
    @OneToMany(cascade = CascadeType.ALL)
    private List<OrderItem> items;
    @OneToOne(cascade = CascadeType.ALL)
    private Delivery delivery;

    public Order(String customer_name, List<OrderItem> items, Delivery delivery) {
        this.customer_name = customer_name;
        this.items = items;
        this.delivery = delivery;
    }

    public Order() {

    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getCustomer_name() {
        return customer_name;
    }

    public void setCustomer_name(String customer_name) {
        this.customer_name = customer_name;
    }

    public List<OrderItem> getItems() {
        return items;
    }

    public void setItems(List<OrderItem> items) {
        this.items = items;
    }

    public Delivery getDelivery() {
        return delivery;
    }

    public void setDelivery(Delivery delivery) {
        this.delivery = delivery;
    }
}