# Lista 3/zad-1/src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /updateStatus:
    put:
      tags:
      - order-controller
      operationId: updateStatus
      parameters:
      - name: orderId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: true
        schema:
          type: string
          enum:
          - CREATED
          - PICKED_UP
          - DELIVERED
      responses:
        "200":
          description: OK
  /orderHistory/updateStatus:
    put:
      tags:
      - order-history-controller
      operationId: updateStatus_1
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /orderHistory/createOrderHistory:
    post:
      tags:
      - order-history-controller
      operationId: createOrderHistory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderHistory'
        required: true
      responses:
        "200":
          description: OK
  /createOrder:
    post:
      tags:
      - order-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        "200":
          description: OK
  /orderHistory/getById:
    get:
      tags:
      - order-history-controller
      operationId: getById
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderHistory'
...
  /orderHistory/deleteById:
    delete:
      tags:
      - order-history-controller
      operationId: deleteById
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
    OrderHistory:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        customerName:
          type: string
        courierName:
          type: string
        deliveryStatus:
          type: string
        productNames:
          type: string
        totalPrice:
          type: number
    Delivery:
      type: object
      properties:
        id:
          type: integer
          format: int64
        courierName:
          type: string
        deliveryStatus:
          type: string
          enum:
          - CREATED
          - PICKED_UP
          - DELIVERED
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customer_name:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        delivery:
          $ref: '#/components/schemas/Delivery'
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          format: int32
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
...